Conflicts:
	src/syscall.cc

+++ src/mapping.cc
@@ -1,10 +1,10 @@
 #include <cstring>
 #include <vector>
 
-#include <elkvm.h>
-#include <elkvm-internal.h>
-#include <mapping.h>
-#include <region.h>
+#include <elkvm/elkvm.h>
+#include <elkvm/elkvm-internal.h>
+#include <elkvm/mapping.h>
+#include <elkvm/region.h>
 
 namespace Elkvm {
   Mapping::Mapping(std::shared_ptr<Region> r,
@@ -37,7 +37,7 @@ namespace Elkvm {
 
   std::shared_ptr<Region> Mapping::move_guest_address(off64_t off) {
     assert(off > 0);
-    assert(off < length);
+    assert((unsigned)off < (unsigned)length);
     addr += off;
     length -= off;
     mapped_pages = pages_from_size(length);
